# docs/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoonFS HTTP API",
    "description": "Static OpenAPI document for the LoonFS v0 HTTP API.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.2.1"
  },
  "paths": {
...
    "/v0/namespaces/{namespace_id}/filesystem/downloads": {
      "post": {
        "tags": [
          "filesystem"
        ],
        "summary": "Begin download",
        "description": "Authorizes one direct read of a file's content object and returns a short-lived presigned GET capability, the resolved revision, and the content reference the client checks the arriving bytes against. `Range` is outside the signature, so one grant serves ranged, resumed, and parallel reads. Deployments that cannot presign answer 501 `not_supported`; the proxied `GET /filesystem/content` route stays available and is capped by `download.max_content_bytes`.",
        "operationId": "begin_download",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Download authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginDownloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace, path, or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Direct download is unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v0/namespaces/{namespace_id}/inodes/{inode_id}/revisions/{revision_no}/downloads": {
      "post": {
        "tags": [
          "inodes"
        ],
        "summary": "Begin download by inode",
        "description": "Authorizes a direct read of one retained inode revision. The request body is `{}` and the response does not include a path.",
        "operationId": "begin_download_by_inode",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inode_id",
            "in": "path",
            "description": "File inode ID",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ino_[1-9][0-9]*$"
            },
            "example": "ino_123"
          },
          {
            "name": "revision_no",
            "in": "path",
            "description": "Revision number",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RevisionNo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginDownloadByInodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Download authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginDownloadByInodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inode ID, revision number, or request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace, inode, or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "408": {
            "$ref": "#/components/responses/DeadlineExceededResponse"
          },
          "409": {
            "description": "Inode is not a file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "410": {
            "description": "Namespace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Direct download is unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbsolutePath": {
        "type": "string",
        "description": "Validated complete absolute namespace path, serialized as a plain string.",
        "example": "/docs/report.txt"
      },
...
      "ApiError": {
        "type": "object",
        "description": "HTTP error body used by LoonFS APIs.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorDetails",
                "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
              }
            ]
          },
          "feature": {
            "type": [
              "string",
              "null"
            ],
            "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "param": {
            "type": [
              "string",
              "null"
            ],
            "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
          }
        }
      },
...
      "BeginDownloadByInodeRequest": {
        "type": "object",
        "description": "Empty request for an inode-addressed download.",
        "additionalProperties": false
      },
      "BeginDownloadByInodeResponse": {
        "type": "object",
        "description": "A short-lived capability to read one inode revision.",
        "required": [
          "namespace_id",
          "inode_id",
          "revision_no",
          "content_ref",
          "access"
        ],
        "properties": {
          "access": {
            "$ref": "#/components/schemas/ObjectTransferAccess",
            "description": "Short-lived provider access without the raw object key."
          },
          "content_ref": {
            "$ref": "#/components/schemas/ContentRef",
            "description": "Content identity, size, and checksum."
          },
          "inode_id": {
            "type": "string",
            "description": "Stable inode ID within a namespace",
            "example": "ino_123",
            "pattern": "^ino_[1-9][0-9]*$"
          },
          "namespace_id": {
            "$ref": "#/components/schemas/NamespaceId",
            "description": "Namespace that was read."
          },
          "revision_no": {
            "$ref": "#/components/schemas/RevisionNo",
            "description": "Revision being read."
          }
        }
      },
      "BeginDownloadRequest": {
        "type": "object",
        "description": "What a client names when it asks to read a file directly.\n\nA path and, optionally, the revision it wants — the same two things the\nproxied content read takes, so a caller switching transports changes\nnothing about what it is asking for.",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "$ref": "#/components/schemas/AbsolutePath",
            "description": "Absolute path of the file to read."
          },
          "revision_no": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RevisionNo",
                "description": "Revision to read, or `None` for the path's current revision."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "BeginDownloadResponse": {
        "type": "object",
        "description": "A short-lived capability to read one file's content object, plus\neverything the reader needs to check what arrives.\n\nThe raw object key is deliberately not here, exactly as it is not in a\n`direct_put` grant: a client learns a URL that expires, not an address it\ncan revisit.\n\nThe grant names one immutable content object, so it does not go stale\nwhen the path moves on. A commit that replaces the file writes a new\nobject and leaves this one alone; what the capability reads is what the\nrequested revision held when the grant was issued, and the reference\nsays which bytes those are.",
        "required": [
          "namespace_id",
          "path",
          "revision_no",
          "content_ref",
          "access"
        ],
        "properties": {
          "access": {
            "$ref": "#/components/schemas/ObjectTransferAccess",
            "description": "Short-lived read capability the client uses without learning the raw object key."
          },
          "content_ref": {
            "$ref": "#/components/schemas/ContentRef",
            "description": "Identity, byte length, and checksum evidence for the object the\ncapability reads. A reader checks the bytes it receives against\n`size_bytes` and recomputes `checksum.algorithm` over the complete\npayload."
          },
          "namespace_id": {
            "$ref": "#/components/schemas/NamespaceId",
            "description": "Namespace that was read."
          },
          "path": {
            "$ref": "#/components/schemas/AbsolutePath",
            "description": "Absolute path as rendered from stored display names."
          },
          "revision_no": {
            "$ref": "#/components/schemas/RevisionNo",
            "description": "Revision the capability reads, resolved from the request."
          }
        }
      },
...
      "ChangeSeq": {
        "type": "integer",
        "format": "int64",
        "description": "Sequence number assigned to a namespace commit. It determines the order in which commits become visible.",
        "maximum": 9007199254740991,
        "minimum": 0
      },
...
      "ContentRef": {
        "type": "object",
        "description": "Pointer to one immutable content object.\n\n`content_id` is identity — *which* object — and the checksum is\nevidence about its bytes. Separating the two is what lets the final\nobject key exist before the first byte is read.\n\nA `ContentRef` is safe to publish only after the referenced bytes are\ndurable in the namespace's content store.",
        "required": [
          "kind",
          "content_id",
          "size_bytes",
          "checksum"
        ],
        "properties": {
          "checksum": {
            "$ref": "#/components/schemas/Checksum",
            "description": "Mandatory checksum over the complete object."
          },
          "content_id": {
            "$ref": "#/components/schemas/ContentId",
            "description": "Immutable identity of the referenced object."
          },
          "kind": {
            "type": "string",
            "description": "Content strategy used by the referenced object."
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Complete byte length of the referenced content.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "ErrorDetails": {
        "type": "object",
        "description": "Optional machine-readable details for an [`ApiError`].\n\nClients make retry decisions from the error code and use these fields for\nrelevant identifiers such as commit ids, writer epochs, and revisions.\nFields may be absent and clients must ignore fields they do not use.",
        "properties": {
          "active_acquired_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Unix milliseconds at which the current epoch's acquirer took it, when\nthe head recorded one. Writer ids are process labels, so two runs on\none machine can share one; the stamp is what tells them apart.",
            "minimum": 0
          },
          "active_deletion_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Deletion generation actually active for the inode."
              }
            ]
          },
          "active_writer": {
            "type": [
              "string",
              "null"
            ],
            "description": "Writer id recorded by the current epoch's acquirer, when the head\nrecorded one."
          },
          "active_writer_epoch": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WriterEpoch",
                "description": "Epoch that currently owns the namespace."
              }
            ]
          },
          "actual_attributes_revision_no": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AttributeRevisionNo",
                "description": "Attribute revision that is actually current for the inode."
              }
            ]
          },
          "actual_head_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Head sequence the namespace was actually at, which is what a caller\nthat still means to delete it retries against."
              }
            ]
          },
          "actual_revision_no": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RevisionNo",
                "description": "Revision that is actually current; absent when the inode has none."
              }
            ]
          },
          "after_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Change-feed cursor the request asked to resume after."
              }
            ]
          },
          "commit_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CommitId",
                "description": "Idempotency key of the commit the error concerns."
              }
            ]
          },
          "committed_fingerprint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Semantic identity of the mutation that already landed under that\ncommit id, from the same receipt as `committed_seq` and present\nexactly when it is. A retry recomputes this value from the request it\njust made — see\n[`put_retry_fingerprint`](crate::put_retry_fingerprint) — and equality\nis what proves the two are the same request."
          },
          "committed_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Sequence at which that commit id already landed. Present when the\nfailure was decided against a durable commit receipt, which is what\nholds the sequence; absent when nothing has committed under the id\nyet and two live requests are simply claiming it at once."
              }
            ]
          },
          "expected_attributes_revision_no": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AttributeRevisionNo",
                "description": "Attribute revision the request expected to be current."
              }
            ]
          },
          "expected_head_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Head sequence a namespace delete required the namespace to still be\nat."
              }
            ]
          },
          "expected_revision_no": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RevisionNo",
                "description": "Revision the request expected to be current."
              }
            ]
          },
          "fenced_epoch": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WriterEpoch",
                "description": "Epoch the failing writer session held when it was displaced."
              }
            ]
          },
          "inode_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "description": "Stable inode ID within a namespace",
                "example": "ino_123",
                "pattern": "^ino_[1-9][0-9]*$"
              }
            ]
          },
          "operation_index": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Position, in the request's operation list, of the operation that\nfailed. A commit applies all of its operations or none of them, so\nthis names the one that stopped the whole request.",
            "minimum": 0
          },
          "requested_deletion_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Deletion generation an undelete asked to recover."
              }
            ]
          },
          "retention_floor_seq": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeSeq",
                "description": "Oldest sequence still promised for incremental replay."
              }
            ]
          }
        }
      },
...
      "NamespaceId": {
        "type": "string",
        "description": "Durable id for one namespace.\n\nA namespace is one filesystem history. This id is not a display name and\nshould not be reused after destruction. Its serialized form is 1 to 128\nlowercase ASCII letters, digits, dots, underscores, or hyphens, starting\nwith a letter or digit; the `loonfs-` prefix is reserved for system use.",
        "example": "demo",
        "pattern": "^[a-z0-9][a-z0-9._-]{0,127}$"
      },
      "ObjectTransferAccess": {
        "oneOf": [
          {
            "type": "object",
            "title": "ObjectTransferAccessPresignedUrl",
            "description": "Short-lived URL plus required headers for one object-store write.",
            "required": [
              "method",
              "url",
              "expires_at_ms",
              "kind"
            ],
            "properties": {
              "expires_at_ms": {
                "type": "integer",
                "format": "int64",
                "description": "Expiration timestamp in Unix milliseconds.",
                "minimum": 0
              },
              "headers": {
                "type": "object",
                "description": "Headers that are covered by the signature and must be sent.",
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "kind": {
                "type": "string",
                "enum": [
                  "presigned_url"
                ]
              },
              "method": {
                "type": "string",
                "description": "HTTP method the client must use."
              },
              "url": {
                "type": "string",
                "description": "Full presigned URL."
              }
            }
          }
        ],
        "description": "Client-facing direct transfer capability."
      },
...
      "RevisionNo": {
        "type": "integer",
        "format": "int64",
        "description": "Revision number for a file's content. It increases whenever the content is replaced or restored.",
        "maximum": 9007199254740991,
        "minimum": 0
      },
...
    },
    "responses": {
      "DeadlineExceededResponse": {
        "description": "The request exceeded `request_deadline_ms`. A timed-out mutation may still complete, so clients must determine its outcome before retrying.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "HTTP error body used by LoonFS APIs.",
              "required": [
                "code",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable reason from the [`ErrorCode`](crate::ErrorCode)\nregistry.\n\nCarried as a string so clients keep working when a newer server\nintroduces a code they do not know; use\n[`ErrorCode::parse`](crate::ErrorCode::parse) for typed access."
                },
                "details": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorDetails",
                      "description": "Structured context for the code, present when the failure carries\nmachine-usable identity (API spec, \"Standard error contract\"). Boxed\nso the rare detailed error does not widen every error-carrying result."
                    }
                  ]
                },
                "feature": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "For `not_supported` errors, the capability-document feature key the\nclient should reconcile against."
                },
                "message": {
                  "type": "string",
                  "description": "Human-readable error message."
                },
                "param": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Identifies the invalid input. Body fields use JSON Pointer paths;\nquery and path parameters use their names; CLI errors use the flag or\nargument as written."
                },
                "request_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Correlation id the server assigned to the failed request; the same\nvalue is sent as the `x-request-id` response header."
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The deployment's `auth_token`, sent as `Authorization: Bearer <token>`. A server configured without a token accepts every request; one configured with a token answers 401 `unauthorized` without it."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}